requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
psycopg2-binary==2.9.9
pymysql==1.1.0
pillow==10.1.0
moviepy==1.0.3
python-dotenv==1.0.0
pydantic-settings==2.0.3
pytz==2023.3
email-validator==2.1.0
httpx==0.25.1
redis==5.0.1
openai>=1.0.0
gunicorn==21.2.0

# AI助手相关依赖
aiohttp==3.8.6  # 异步HTTP客户端，用于调用阿里云API
requests==2.31.0  # HTTP请求库，备用选项 